+++ chalkbox/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.2.0"
+__version__ = "2.0.0"
 
 from .components.alert import Alert
 from .components.align import Align